Add Input Validation (#68)
This pull requests adds the necessary structure to validate inputs for each module as a first step, before the scenario inputs are created. We add a new validate_inputs.py script that looks very similar to get_scenario_inputs - it also loads the inputs from database - but rather than writing the inputs to .tab files, the inputs will be subject to input validation tests. 
The results of the input validation are written to the database and the scenario validation status is updated based on the results of the input validation. 

Changes: 
 - In get_scenario_inputs.py, rename get_inputs_from_database to write_model_inputs to 
   distinguish getting inputs through SQL queries (also used in input validation) from writing
   those inputs to .tab files.  
 - Add a validate_inpputs.py script that goes through each module, gets the inputs from 
   database, and validates them (note: no inputs will be written to .tab files). 
 - In the individual modules, split out get_inputs_from_database into 
    get_inputs_from_database (load the inputs) and write_model_inputs (write the inputs to 
    .tab files). When calling 'write_model_inputs' in get_scenario_inputs.py, each module will
    first get the inputs from database, and then write them to tab files (2 step process). 
 - add a validate_inputs function in each module which will get the inputs from database 
   and validate the inputs (2 step process). 
 - add an option to determine modules based on the database rather than a features.csv file.
   This is necessary for the input validation which does not depend on any local .tab of .csv 
   files 
 - Add database tables that track the input validation result and the validation status