View input data from the UI (#74)
This implements the ability to view DB input data tables from the 
scenario-detail screen of the UI. A view-data service is used to 
subscribe to each table's API, now implemented on the server side.
All tables are currently on the same Angular route (shown and hidden 
with an *ngIf), which does seem a bit slow with a large-ish database. 
In addition, all table data, not just the 'subscenario' is shown, 
which is perhaps more appropriate for the scenario-new screen.