Read DataFrame and columns into one step (#87)
Context: we read in inputs into dataframes by first executing a cursor
query and then converting the query results into a DataFrame in 2 steps:
1. create the DataFrame from the raw query results (fetchall)
2. set the columns to the query result column values (description)

This 2 step approach would throw an error when trying to set columns
to an empty dataframe. With this update, an empty query will still
create a dataframe with the proper columns, but it will simply have no
entries. This makes validating the inputs much easier (no need to check
with a bunch of if statements).

This update also fixes how to deal when there's no availability
subscenario specified.