Change scenario-new form control keys
The keys are now a string of the format 'ui_table' based
on the values from the ui_scenario_detail_table_row_metadata table.
The server and Angular only need to be aware of that key, based on
which other metadata for creating/editing scenario can be pulled from
the table.