Add popper.js and jquery packages
These are bootstrap dependencies that are not actually needed since
we are using ng-bootstrap, which replaces the two. However, an error
is thrown on the initial npm install if the two are not present
("Error: Unresolved node modules: jquery, popper.js"), so I'm adding
them as dev dependencies as a temporary fix.