Embed results plots in the UI (#115)
A structure is now in place for embedding plots in the UI, including 
for selecting subsets of data to display a particular plot (e.g. the 
dispatch plot for a particular load zone and horizon).

The dispatch plot with the selected load zone and horizon is the only 
available plot. Currently, clicking the dispatch plot button again 
displays the next plot without removing the previous one, so any 
number of dispatch plots can be displayed at the same time.