Settings status in UI home component (#120)
Settings status can be 'not_set', 'set' or 'restart required' (when
current setting does not match requested setting).

For the purpose, we now keep track of the 'requested' and 'current' 
settings. Both are stored in electron-json-storage, with the 'current'
set to the 'requested' on Electron launch (this is the only time the 
'current' setting is modified).

A Behavior subject in a new settings service has a default value of 
'set' for each setting. If the setting is changed (new 'requested' 
setting, stored by Electron) and it does not match the 'current' 
setting, then the status is changed to 'restart required' for the 
respective setting (if the 'requested' is changed back to match the 
'current' setting, the status will be 'ok' again).

Electron now checks for whether required settings are stored before 
launching the server (server is not launched if not) and, if not, 
creates the keys for those settings, but sets them to null. The home
component renderer now also communicates with Electron to get the 
stored settings. If it receives 'requested' setting values of null, 
it sets the status for the setting to 'not set' -- this can only 
happen when the UI is first launched and settings have not yet been 
set, as we currently do not allow settings to be changed back to null
after being set.