Features table in scenario-new standard API
The features table is now built with the same API as the rest of the
tables in scenario-new. It requires one extra key -- the name of the
column in the scenarios_view database view -- to get the starting
values for the checkboxes to work. To avoid the need for the extra key,
we should change the starting values API to use the ui_table_name_in_db
and ui_table_row_name_in_db instead.