Base edit-scenario API on ui_scenario_detail tables
Instead of the scenarios_view column names, the edit-scenario API is
now based on the combined 'ui_table$ui_table_row' values from the
database. The edit-scenario starting value functionality for the
feature checkboxes can now use that instead of the scenarios_view
column names, so small changes removing the rowDBViewName field are
also implemented for the scenario-new API in this commit.