Format the y-axis of the viz plots
Formats the axis to write out the full number with a thosuands seperator
(no scientific formatting).

Note that a more advanced solution could be to look at the maximum
value and change the units of the y-axis accordingly.
(e.g. if less than 1 MW, change to kW; if more than 1000 Mw, change
to GW etc.).
For now I think always writing it out fully and using a thousands
separator seems adequate, but open for suggestions.

Closes #126