Add energy by period plot (#132)
This plots the energy by period for a select zone and stage.

The stage defaults to 1 which should be the default if you're not
modeling multiple stages. We also assume that spinup or lookahead
timepoints are identified by having anything but Null in the
spinup_or_lookahead timepoint (we might want to document/enforce this
better and or change the convention to something that we think is better)

The energy is calculated by multiplying each technology's power output
in each timepoint by that timepoint's number_of_hours_in_timepoint and
that timepoint's corresponding horizon's horizon_weight, and aggregating
it by technology and period.

We assume that the timepoints within each period have horizon_weights
and timepoint durations that sum up to a logical number that is the
same for each period (generally 8760 if periods are 1 year). The input
validation should warn the user if that is not the case.

Note: we might want to consider adding curtailment and imports to this
chart.