Add cost by period plot (#135)
This plots the cost by period for a select scenario/zone/stage, broken
out by the different cost categories.

The stage defaults to 1 which should be the default if you're not
modeling multiple stages.

Except for the capacity cost, the cost is calculated by multiplying each
projects's cost in each timepoint by that timepoint's
number_of_hours_in_timepoint and that timepoint's corresponding
horizon's horizon_weight, and aggregating it by technology and period.
This is done for each cost category. For startup and shutdown costs, we
don't weigh by the timepoint duration since the costs do not scale with
the duration.

We assume that the timepoints within each period have horizon_weights
and timepoint durations that sum up to a logical number that is the
same for each period (generally 8760 if periods are 1 year). The input
validation should warn the user if that is not the case.

Notes:
 - We need to figure out how to allocate hurdle rate costs to zones
 - We need to think about how to deal with performance issues
   if we were to compare multiple scenarios as well (not slicing for
   scenario results in queries taking ~20s).
 - More generally, we need to think about how to deal with the other
   dimensions. Sometimes we might want to compare costs/energy/capacity
   across scenarios for a given period/zone, or across zones for a given
   period/scenario. There might be a way to generalize this so we don't
   have to write 3 different plot functions for that. The general case
   is that you have a result that is indexed by 3 things, so you want
   to keep 2 things fixed while varying the last one.