Refactor UI add-scenario functionality
The updated functionality for adding scenarios from the UI via
./ui/server/db_ops/add_scenario.py takes advantage of the new
create_scenario function from the ./db/utilities/scenario.py module.
The client message is converted to a dictionary that is given as an
argument to that function (we can create the dictionary keys from the
form keys) instead of using the create_scenario_all_args function that
requires all columns in the scenarios table to be specified
individually. In addition, we can now take advantage of the greater
flexibility of scenario.create_scenario, namely that missing columns
and/or None values are allowed. This simplifies treatment in the
situations when we have a UI that hasn't exposed all columns or the
user has not specified values for certain subscenarios. In the latter
case, we want the scenario-creation to still go through, as our
approach is to catch potential issues with the validation functionality.