Pass scenario ID to view-data
If 0 is passed, all data from a table will be shown. The scenario-new
component passes 0 when the user requests to view a table.

The scenario-detail component passes the scenario ID of the scenario
that the user is viewing and clicking a 'View Data' button shows only
the data from the subscenario_id selected for the scenario.