Blank name on populating new from existing scenario
Previously, the scenario name was populated with the name of the
existing scenario providing the starting values. To prevent a situation
in which the user accidentally edits an existing scenario instead of
creating a new one, the scenario name field is now blank, even when
populating from an existing scenario.

More detailed treatment is still needed for how to handle editing
scenarios, a scenario name that already exists being specified, etc.