Use newline="" when opening files for writing (#150)
On Windows, CSV/TAB files were written with an extra blank line in 
between lines. This was not only annoying, but was actually breaking 
get_inputs_from_database.py, as modules needing to open and append to 
files would find a blank line when iterating over rows and throw an 
index error.

Adding newline="" is the suggested solution on Python 3.