Add project operations plot
This plots the operations for a given scenario/project/stage/period. It
assumes that the project is one of the operational types with
commitment decisions (ValueError if it isn't).

If there are reserves, they will be shown as shaded areas above/below
the "Power" line.

The user can optionally specify a starting and ending horizon.