Routing-based scenario-new UI component (#158)
This changes how the 'starting values' are populated in scenario-new.
Previously, we were using the scenario-edit service under
scenario-detail and a BehaviorSubject that was used to pass the
starting values when editing a scenario or populating a new scenario
based on an existing one. The new approach is to use routes to know
which starting values to use. For the purpose, we add a routing :id to
the scenario-new route. The scenario-edit service modules are deleted.

When the user clicks on "New Scenario" (now available both from the
scenarios list view and the navigation bar), we pass the route with
:id=0. This route is linked with empty starting values. We also now
pass some 'navigation extras,' one for whether to 'hide' the starting
scenario name and one for whether to disable the scenario name input
field. In the route :id=0, both are set to false. The empty starting
values are now created based on the row-identifiers from the API
instead of hard-coded.

If the user decides to populate the new scenario based on an existing
one, we will now route to that existing scenario's ID and use the
scenario-detail service to get the starting values. In this case, we
also pass true for the hideScenarioName navigation extra, which sets
the scenario name to blank, helping to avoid a situation in which the
user accidentally overwrites an existing scenario.

If the user decides to edit an existing scenario, we will now route
to that existing scenario's ID and use the scenario-detail service to
get the starting values. In this case, we also pass true for the 
inactiveScenarioName navigation extra, which disables the scenario 
name field, helping to avoid a situation in which the user
accidentally creates a new scenario instead of editing the existing
one they selected. We also use the inactiveScenarioName flag to
change the heading to 'Edit Scenario' (instead of 'New Scenario').

Note that we must get the navigation extras from the browser history,
as they are only available during navigation.