Refactor UI scenario-results view (#164)
The UI scenario-results view now has a flexible structure that allows
us to use database tables to create results table and plot components
in scenario-results.

The two tables are defined in the database schema and are called
ui_scenario_results_table_metadata and
ui_scenario_results_plot_metadata respectively. The user
can simply change the value of the 'include' column of these database
tables to decide whether the results table or plot specified in the
row will be included in the scenario-results view. Data for all 
available results tables and all available results plots is loaded 
automatically on database creation from the db/data directory with
include=1.

For the results tables, the user must specify the table name and UI
caption in ui_scenario_results_table_metadata.

For the plots, the user must specify the name of the plot-generating
script, the plot caption in the UI, and whether or not a form control
is needed for the plot via a boolean value in the respective column
(e.g. a drop down menu for load zones is included in the plot form
by specifying 1 in the load_zone_form_control column) in 
ui_scenario_results_plot_metadata.