Refactor view-data component (#166)
The view-data component is renamed scenario-inputs and refactored, so
that we get the right right table in the view from the route and the
NavigationExtras on the route (instead of with an *ngIf). For 
scenario-inputs the navigation extras are the table and row names from
scenario detail, which are then matched to the scenario detail metadata 
database tables.

A new view-data component is created for other table views that are
not part of the scenario-detail structure. This currently contains
only the validation status. This component is also based on the route:
scenario ID and navigation extras, which is this case is the database
table name.