Add scenario-comparison to UI (#169)
This adds a scenario-comparison-select component and a 
scenario-comparison component that allow the user to easily
compare scenario inputs in the same view. In the former, the user can
choose a base scenario and select any number of other scenarios to 
compare with the base. In the scenario-comparison component, a view is
generated that follows the scenario-detail structure and highlights
the settings in each scenario being compared that are different from
those of the base scenario. The user can also switch to the
scenario-inputs view from here.