Move reserve types to database
Previously these were hard-coded in the python code which is not
advisable.

Note: the reserve types still show up hard-coded in some places such
as the operations plotting modules. We could think of adding a
third column to the reserve_types table that lists the cleaner name used
for plotting.