Add subproblem/stage index/arg to relevant plots
The capacity, carbon, and rps plots were all based on results that
are indexed by period, subproblem, and stage but they weren't slicing
out a particular subproblem/stage. This update adds a subproblem and
stage argument to each of these plots to slice out the appropriate
results.

When no subproblem/stage is provided, the default is set to 1 which is
assumed to be the default subproblem/stage when running the model in
capacity expansion mode.

Note: we might want to disallow users to combine RPS/carbon cap/new
build functionality with multi subproblem/stage functionality anyway
since usually this will be meaingless
(see issues #99, #173, and #175).

Closes #173