Adjust legend sizing dynamically
Add a function that checks how many legend entries there are and
how long they are to determine the appropriate font size dynamically.

The issue with the current approach is that the relationship between
font size, graph height, and number of legend entries appears to be
non-linear. The current linear approach results in a fairly "short"
legend when charts are tall (vs. the default) and still results in a
legend that is too tall (overflows) when the chart is very short.
You can see this by changing the chart size from 500 to 300 and looking
at both charts with and without the resizing function.