Base expected datatypes on database columns (#174)
Previously these datatypes were hard-coded in the python code while
the expected datatypes were suggested in the sqlite database table
structures all along.

This update checks the expected datatypes using the PRAGMA command
and uses that to create a dictionary with the expected datatypes by column 
name. 

To get the datatype category (numeric or string) based on the specified
column datatypes, rather than looking for exact datatype descriptions, we 
use the 5 sqlite rules that are used to determine the affinity of a column 
based on the declared type of a column. See: 
https://www.sqlite.org/datatype3.html