Flexible balancing horizons (#200)
This introduces additional flexibility to the GridPath 'horizons' by
introducing the concept of 'balancing type'. Horizons can be thought
of as a way to organize timepoints into groups, i.e. the same timepoint
can now belong to multiple horizons, e.g. a day of the year, a week of
the year, and the year. Horizons are still unique and they now have an
extra parameter called 'balancing_type_horizon' that defines their
length. The horizon weight param is removed. Instead, the parameter
timepoint_weight is introduced, moving 'weighting' to the timepoint
level. All projects must now also be assigned a balancing type (the
parameter is called balancing_type_project), which determines the
timeframe over which they are balanced.