Add New Binary Build Generator capacity type (#208)
This capacity type is similar to the new_build_generator capacity
type, except that it makes the new build decision binary rather
than continuous. This means that the model can decide to either
build the project at a certain size, specified by the new input 
parameter "binary_build_size_mw", or not build it at all.

We add a new input table "inputs_project_binary_build_size" to 
the database to accommodate the new input variable, as well
as functionality to create different subscenarios for different 
binary build size settings. 

Outputs are saved in a new databased table called 
"results_project_capacity_new_binary_build_generator",
as well as in the existing table "results_project_capacity_all"

For testing, we add a new test scenario in the examples called 
"test_new_binary_solar" and we slightly change the 
"test_new_solar" example so the comparison with
this new binary example makes more sense.

The test data for the unittest is also updated to include an
example of this new capacity type. 

Addresses one of the 3 updates in #21.