Improved user-facing scenario-process-handling (#222)
This introduces several improvements to how the user can manage
scenario processes.

When a scenario process is launched, its process ID is now recorded in
the database. We can then use this process ID to stop a scenario run if
the user requests so. 

We also introduce 'action' buttons in the scenario-detail view that are
based on the run status, e.g. the 'stop run' button appears when a
scenario is running, the 'run scenario' button is available when a
scenario has not been run or a run error has been encountered, and 
the 'clear scenario results' button appears when a scenario run has 
completed. A new run status 'run_stopped' is introduced for the
situation when a user deliberately stops a scenario (different from a
run_error). The goal for the status-based action buttons is to allow
only certain actions based on the run status (e.g. a scenario can only
be run if it's not already running or complete; otherwise, it needs to
be stopped or cleared respectively; a scenario cannot be stopped if
it's not running; a scenario cannot be cleared if it's not complete,
and so on). To prevent a user from launching a second or third, etc.
scenario run process after the initial one in the few seconds before
the run status updates to 'running' and the action button changes, we 
also disable the 'Run Scenario' button on click (it is reenabled on 
refresh, and when the scenario run is stopped or the scenario is
cleared.

A 'delete scenario' button is also implemented that allows the user to
fully remove the scenario from the database (i.e. not just the results,
but also remove from the scenarios table).

Several user warnings are introduced, including for clearing a
scenario, deleting a scenario, and quitting the app (which would stop
all running scenarios).

A note: on Windows, the scenario process is now started with
shell=False.