New Binary Build Storage Capacity Type (#216)
This capacity type is similar to the new_build_storage capacity
type, except that it makes the new build decision binary rather
than continuous. This means that the model can decide to either
build the project at a certain size, specified by the new input
parameters "binary_build_size_mw" and "binary_build_size_mwh",
or not build it at all. These new input parameters were added in
#208 as part of the new input table
"inputs_project_binary_build_size".

Note that the storage duration sizing in this framework is not
endogenous but rather specified by the ratio of the MWh
and MW of the specified build size.

Outputs are saved in a new databased table called
"results_project_capacity_new_binary_build_storage",
as well as in the existing table "results_project_capacity_all"

For testing, we add a new test scenario in the examples called
"test_new_binary_build_storage".

The test data for the unittest is also updated to include an
example of this new capacity type.

Addresses one of the issues in #21.
This PR is pretty much exactly the same as PR #208 which
introduced the new capacity type "new_binary_build_generator",
but for storage.