UI plots accept default values for stage/subproblem
This adds subproblem as a dropdown for the UI plots. If the user does
not specify a subproblem and/or stage, the default values, as specified
in the plot script arguments, are used as respective flag is not
passed. This also makes it possible to omit the dropdowns when we only
have a single subproblem and/or a single stage, as in that case the
script defaults are also used.