Change hydro operational units
This PR changes the units to fraction of installed capacity rather than
absolute amounts. This allows users to combine new build capacity types
with hydro operational types, since we can multiply the fraction with
the capacity expression, which can be a decision variable in the case
of new build resources.

Note: this requires an update to the port_irp_data_to_gridpath script
to normalize the hydro inputs to existing capacity.

Closes #48