Make 'soft' load-balance constraint optional (#227)
This introduces boolean params for allowing or disallowing load-balance 
constraint violations (overgeneration and unserved energy). These
parameters get get multiplied by the Overgeneration and Unserved_Energy 
variables to create expressions, which then get added as components to
the load-balance constraint instead of the variables themselves. When 
overgen/unserved energy is disallowed, these expressions are simply 0,
so the constraint is hard.

The definition of the load-zone-level params are moved to the
geography.load_zones module from the system.load_balance.load_balance
module.