Add additional sets to DC OPF Module (#235)
Add sets to DC OPF module so you can deal with a transmission network that changes across modeling periods. The cycles, and cycle direction parameters are now also indexed by the period. 

Note: the unittest now checks for both cycle direction to be able to deal with the random cycle directions used in the networkx package (because it uses set.pop()).