refactor determine_relevant_timepoints_startup
Can simply can determine_relevant_timepoints twice and take the
difference between the resulting set.

Note that there is a slight change here with the previous version:
If we are within TSU,s+1 hours (but > TSU,s hours) from the start of the
linear boundary horizon, we will now actually enforce the constraint
whereas before the constraint would be skipped as soon as you'd be
within tSU,s+1 hours from the start.