Add an argument group for required parsed args (#299)
This way the --help message will be more clear and won't list the
required arguments under optional arguments, which is the default
for argument flags (--arg or -a), whether they are required or not.

Note: to show the required arguments before the optional arguments
we rely on a hacky fixed that accesses and changes a private attributes 
of the parser object (it reverses the list order). 

This addresses some of the issues raised in #111.