Clean up after processes that did not exit cleanly (#309)
In some cases, if a scenario processes crashes (e.g. due to a system
problem), it will not clean up after itself, so scenarios will appear
as running, even though they really crashed. If the user tries to
'stop' such a scenario, we now catch the fact that the process ID does
not exist, update the scenario run status to 'run error' and clear any
scenario results.