Custom plotting colors and order (#312)
This PR allows users to optionally specify custom plotting colors and orders for technologies. The plotting scripts read a new database table (viz_technologies) with technologies and their specified color and order and uses that mapping in the plots. We provide a set of default inputs for viz_technologies when the database is created, but users can change this table through the UI (to be implemented).

If there is no color specified (either technology is not in table or color is NULL), a standard color palette (cividis) is used for all unspecified technologies.

If there is no order specified (either technology is not in table or plotting_order is NULL) the technology will be plotted after any technologies that have an order provided, sorted alphabetically with any other technologies that have no order specified.

This update also removes hard-coding of technologies in the dispatch plot.

See also #112.