Format undefined subscenarios in SubScenarios class (#322)
In the SubScenarios class of gridpath.auxiliary.scenario_chars, format
undefined scenario so that we get NULL rather than None. This is
better since the subscenario elements of this class are used to format
the various get-inputs queries, which can break unless we use NULL when
a subscenario is not defined.