Startup and Shutdown Trajectories (#297)
This PR adds startup and shutdown trajectories to the gen_commit_bin and gen_commit_lin operational type modules. The inputs are exactly the same as before, but now the unit will take several timepoints to start up (shut down) if the startup_plus_ramp_up_rate (shutdown_plus_ramp_down_rate) is small enough.  

The functionality works by adding 2 additional decision variables to represent the power during startup and shutdown (when the unit is below its minimum stable level) namely Pstarting and Pstopping. These variables are constrained in such a way that they obey the startup and shutdown ramp rate limits and are zero when the unit is actually operating above its minimum stable level. 

The updates are implemented for both the gen_commit_bin and gen_commit_lin module. For the latter, all binary variables are made continuous which linearizes the problem and therefore makes it a lot faster to solve. It should be noted that this could lead to some weird behavior with a fraction of the unit starting up while another fraction is shutting down in the same timepoint, so this approach should be used with caution. 

Other changes:
 - new test example with a slow startup and a fast shutdown: test_startup_shutdown_rates
 - input validation to make sure we startup and shutdown inputs are compatible with the other inputs.