Documentation re-org of the gen_commit_cap module (#323)
The new structure for the module documentation is as follows:
1. The module has a high-level description of what it does in a
multi-line docstring at the top level of the module. This is pulled
automatically into the 'high-level' GridPath documentation.
2. The add_module_specific_components function of the module has a
description of the components added to the Pyomo model that are pulled
into the 'advanced' GridPath documentation. There tables for sets,
required and optional parameters, variables, and constraints defined
in the module with short descriptions. The params and variable tables
also include the 'within' and, if any, the 'default' for the
respective component.
3. Constraint rules are taken out of the add_module_specific_components
function and included at the top-level of the module below that
function instead under a header.
4. We now also have a header for the module methods to distinguish
those from the constraint rules.

This commit also includes changes to some Pyomo component names and
various clean-up in the gen_commit_cap operational module, but the
documentation and polishing are not complete yet.