UI run queue (#411)
Implement a basic run queue in the UI. Scenarios can be added to the
queue instead of run immediately. For the purpose, we now have an
additional process that is launched by the server whenever a scenario
is added to the queue; the process exits when the queue is empty. This
process also continuously checks if the server is available and quits
if it is not (instead of using signals from the server process to shut
it down). A new script entry point is created for the queue manager
process (gridpath_run_queue_manager).

Scenarios can be removed from the queue but there is currently no way
to change the order of scenarios in the queue.