Change horizon indexing structure (#438)
Start with a two-dimensional set called BALANCING_TYPE_HORIZONS
instead of a unique HORIZONS set with the balancing_type as param.
This necessitates some indexing changes of the various derived params
(e.g. first and last timepoint of balancing-type-horizon, boundary of
balancing-type-horizon). I added the balancing type to the primary key
of the horizons table in the database, but other than that this is a
silent update.