Fix DB utils for heat rate curves
The way the file structure and for loops are set up in
load_project_operational_chars does not allow for different
heat rate curves subscenarios for one project (the subscenario gets
overwritten by latest subscenario in the latest project outer loop).

This update provides a quick fix to deal with this. A longer term fix
would be to further clean this up similar to the temporal PR (#441)
and/or to create csv input files separately for each project-subscenario
combination, similar to the variable profiles and hydro inputs.
(this would require adding the project column to the subscenario data
csv input).

Closes #453