Hot-cold startup costs (#454)
This PR makes startup costs dependent on startup type, similar to how startup ramp rates were made dependent on the startup type in #440.

We add the startup_cost_per_mw column to the startup_chars input, and the gen_commit_lin and gen_commit_bin operational type will read in the startup costs by startup type from this table.

Note that gen_commit_cap operational types will still only have one startup cost, and the startup cost will be read in from the project_operational_chars table. I.e. the same input can be found in different places, depending on the operational type. This is not ideal and should ideally be changed in a future update.

This PR also includes a fix to the startup_car DB utilities, which weren't properly working for multiple subscenarios per project (see #453 for a similar issue with the heat rates). This was necessary to pass the test_examples in the unittests.

Closes #446