Remove subscenario files from CSV-to-DB (#467)
Instead of relying on a subscenario file that lists the subscenarios
and where their data is located -- and that must be maintained -- the
scanner function now looks for CSV files in the respective subscenario
folder and expects that the names of these files will have a certain
structure to indicate the ID and name of the subscenario, and contain
the data for the subscenario. The subscenario ID does not need to be
included anywhere other than the filename. A description of the
subscenario can be provided in a TXT file with the same name as the
CSV file for the subscenario (if no such TXT file the description is
an empty string).

A check is implemented to warn if subscenario IDs are not unique.