Simplify endogenous availability module (#476)
The constraints in the endogenous availability module were somewhat
redundant, could easily result in infeasible problems, and appeared
to be resulting in MIP problems that were very difficult to solve for
larger cases. I am removing the 'max' constraints here to simplify
those, as the 'min' ones only -- provided the inputs are well
thought-through -- produce a meaningful solution.