Add missing print statements for db utilities
With #467 and #441 the print statements were removed and the
"quiet" argument had become redundant. This commit prints the files
being processed again, when the "quiet" argument is False. An
alternative solution would be to get rid of the "quiet" argument
entirely and never print anything (note that the solver options and
scenario db utils are still using the "quiet" argument regardless of
this commit).

This commit also removes the no-longer-used function
"csv_read_temporal_data".