Make unit size the same in all examples
Some examples had a unit size of 1 MW instead of 6 MW. This is not
consistent with the heat rate curves inputs which assume a unit size of
6 MW and as a result it made the fuel burn much larger than expected,
due to the intercept term being much larger (it is multiplied by the
number of units online, which is affected by the unit size; note that
the slope is unaffected).

The unittests are updated so that they still pass.

This also removes a redundant unit size specification for a
gen_commit_bin generator.