Allow no fuels (#499)
This removes the fuels optional feature and allows all projects not
to have fuels/heat rates specified (regardless of operational type).
The modules that were associated with the fuels feature are now
loaded by default, but they check whether scenario projects have
fuels, get the inputs for those fuels only, and initialize the fuel
data. If the input files for fuel characteristics and fuel prices are
empty, the data are not loaded and the components not initialized.